In the beginning of each script, unload all packages.
It may seem redundant, but when we are interactively running scripts for
debugging purposes, it's common to have some packages that are already loaded
in the session. If a script relies on such packages but doesn't declare them,
it will not work when run from scratch using Rscript, although it will run
if we execute line by line in the current session.